Converted to big decimal

+++ com/acme/account/Account.java
@@ -1,15 +1,19 @@
 package com.acme.account;
 
+import java.math.BigDecimal;
+
+// Updated Double to BigDecimal
+
 public interface Account {
 
 	//Deposit money into the account
-    public void deposit(double amount, TransactionType type);
+    public void deposit(BigDecimal amount, TransactionType type);
 
     //Withdraw money from the account
-    public void withdraw(double amount, TransactionType type);
+    public void withdraw(BigDecimal amount, TransactionType type);
 
     //Get current account balance
-    public double getBalance();
+    public BigDecimal getBalance();
     
     //List transaction history
     public String getHistory();
